pyproject.toml
[tool.poetry]
name = "z3-target-bigquery"
version = "0.5.6"
description = "z3-target-bigquery is a Singer target for BigQuery. It supports storage write, GCS, streaming, and batch load methods. Built with the Meltano SDK."
authors = ["Alex Butler <butler.alex2010@gmail.com>"]
keywords = ["ELT", "BigQuery"]
license = "MIT"
include = ["target_bigquery/*.py", "README.md", "LICENSE"]
homepage = "https://github.com/z3z1ma/target-bigquery"
repository = "https://github.com/z3z1ma/target-bigquery"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
packages = [
    { include = "target_bigquery/*.py" },
    { include = "README.md" },
    { include = "LICENSE" },
]

[tool.poetry.dependencies]
python = "<3.11,>=3.8"
requests = ">=2.25.1"
singer-sdk = ">=0.16.0"
google-cloud-bigquery = { version = ">=3.4.1,<4", extras = ["bqstorage"] }
orjson = ">=3.7.2,<4"
tenacity = ">=8.0.1,<9"
grpcio-tools = ">=1.51.1,<2"
google-cloud-storage = ">=2.7.0,<3"
grpcio-status = { extras = ["grpc"], version = ">=1.51.1,<2" }

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = "target_bigquery"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
target-bigquery = 'target_bigquery.target:TargetBigQuery.cli'
